scripts/pinecone_interaction/preprocessing/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0
tarfile>=0.9.0
gzip>=0.0.1
pylatexenc>=2.8.0
pinecone-client
tenacity
pinecone-plugin-assistant